generate: Move AST reading functionality to 'common'.
