+++ internal/generate/listpages/main.go
@@ -15,9 +15,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform-provider-aws/internal/generate/common"
-	tfslices "github.com/hashicorp/terraform-provider-aws/internal/slices"
 	"github.com/hashicorp/terraform-provider-aws/names/data"
-	"golang.org/x/tools/go/packages"
 )
 
 var (
@@ -80,7 +78,7 @@ func main() {
 	awsPkg := service.GoPackageName()
 
 	sourcePackage := fmt.Sprintf("github.com/aws/aws-sdk-go-v2/service/%[1]s", awsPkg)
-	pkg, err := loadPackage(sourcePackage)
+	pkg, err := common.LoadPackage(sourcePackage)
 	if err != nil {
 		g.Fatalf("parsing package (%s): %s", sourcePackage, err)
 	}
@@ -102,7 +100,7 @@ func main() {
 	functions := strings.Split(*listOps, ",")
 	slices.Sort(functions)
 	for _, functionName := range functions {
-		function := pkg.findFunction(functionName)
+		function := pkg.FindFunction(functionName)
 		if err != nil {
 			g.Fatalf("function (%q) not found", functionName)
 		}
@@ -148,72 +146,6 @@ func main() {
 	if err := d.Write(); err != nil {
 		g.Fatalf("generating file (%s): %s", filename, err)
 	}
-}
-
-type PackageFile struct {
-	file *ast.File
-}
-
-type Package struct {
-	name  string
-	files []*PackageFile
-}
-
-func loadPackage(sourcePackage string) (*Package, error) {
-	cfg := &packages.Config{
-		Mode: packages.NeedName | packages.NeedTypes | packages.NeedSyntax,
-	}
-	pkgs, err := packages.Load(cfg, sourcePackage)
-	if err != nil {
-		return nil, fmt.Errorf("loading %s: %w", sourcePackage, err)
-	}
-	if len(pkgs) != 1 {
-		return nil, fmt.Errorf("%d packages found", len(pkgs))
-	}
-	pkg := pkgs[0]
-
-	return &Package{
-		name: pkg.Name,
-		files: tfslices.ApplyToAll(pkg.Syntax, func(file *ast.File) *PackageFile {
-			return &PackageFile{
-				file: file,
-			}
-		}),
-	}, nil
-}
-
-func (pkg *Package) findFunction(functionName string) *Function {
-	for _, file := range pkg.files {
-		if file.file != nil {
-			for _, decl := range file.file.Decls {
-				if funcDecl, ok := decl.(*ast.FuncDecl); ok {
-					if funcDecl.Name.Name == functionName {
-						return &Function{
-							funcDecl: funcDecl,
-						}
-					}
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-type Function struct {
-	funcDecl *ast.FuncDecl
-}
-
-func (function *Function) Name() string {
-	return function.funcDecl.Name.Name
-}
-
-func (function *Function) Params() []*ast.Field {
-	return function.funcDecl.Type.Params.List
-}
-
-func (function *Function) Results() []*ast.Field {
-	return function.funcDecl.Type.Results.List
 }
 
 func identifierName(expr ast.Expr) string {
